Toggle button styled component migration
