+++ src/index.tsx
@@ -39,7 +39,7 @@ export { default as Text, TextDisplayType, TextTagType } from 'text'
 export { default as TextField, inputTypes, inputModes } from 'textField'
 export { default as TimePicker } from 'timePicker'
 export { default as Title } from 'title'
-export { default as ToggleButton } from 'toggleButton'
+export { default as ToggleButton, ToggleButtonStatus } from 'toggleButton'
 export { default as TopBar } from 'topBar'
 export { default as Transitions } from 'transitions'
 export { default as TripCard } from 'tripCard'
