+++ src/toggleButton/story.tsx
@@ -4,7 +4,7 @@ import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
 import { withKnobs, text, boolean, select } from '@storybook/addon-knobs'
 
-import ToggleButton from 'toggleButton'
+import ToggleButton, { ToggleButtonStatus } from '.'
 
 const stories = storiesOf('ToggleButton', module)
 stories.addDecorator(withKnobs)
@@ -17,6 +17,6 @@ stories.add('ToggleButton', () => (
     disabled={boolean('Disabled', false)}
     label={text('Label', 'Label')}
     sublabel={text('Sub label', 'Sublabel')}
-    status={select('status', ToggleButton.STATUS, ToggleButton.STATUS.DEFAULT)}
+    status={select('status', ToggleButtonStatus, ToggleButtonStatus.DEFAULT)}
   />
 ))
