--- src/toggleButton/style.ts
@@ -1,22 +1,0 @@
-import css from 'styled-jsx/css'
-
-export default css`
-  .kirk-toggle-button {
-    display: block;
-    margin: 0;
-    padding: 0;
-    background: none;
-    width: 100%;
-    border: 0;
-    text-align: left;
-    font-family: inherit;
-    cursor: pointer;
-    user-select: none;
-    overflow: hidden;
-  }
-
-  .kirk-toggle-button:focus,
-  .kirk-toggle-button:active {
-    outline: 0;
-  }
-`
